LayoutTests/fast/scrolling/scroll-animator-overlay-scrollbars-hovered-expected.txt

CONSOLE MESSAGE: line 17: MainFrameView: mouseEnteredContentArea
CONSOLE MESSAGE: line 17: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 18: FrameView: mouseEnteredHorizontalScrollbar
CONSOLE MESSAGE: line 18: FrameView: mouseEnteredContentArea
CONSOLE MESSAGE: line 18: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 19: FrameView: mouseExitedHorizontalScrollbar
CONSOLE MESSAGE: line 19: FrameView: mouseExitedContentArea
CONSOLE MESSAGE: line 19: MainFrameView: mouseMovedInContentArea
PASS successfullyParsed is true

TEST COMPLETE
Test for https://bugs.webkit.org/show_bug.cgi?id=153304.
